test(files): new files for mock tests
